V5.49.3 (#1539)
* Update Expression Parser (#1505)

* bump version for next beta

* Merge v5 updates into beta (#1508)

* Fixed expressionish

* Update expressionish

* Bump version

* Fix $target

* Fix $args

* fix HTTP GET requests

* Update expressionish

* Bump Version

* Add suffix write mode to file write effect

* Register dummy $if variable

* fix examples

* Implement $rollDice var, refactor Dice effect

* support markdown in variable descriptions

* Add $runEffect variable

* Fix default wager for heist game

* Add default wager to the !spin game too

* Bump beta version

* Fix date example

* fix $rollDice

* Update request HTTP requests to axios

* Add support for 7tv emotes

* Remove Mixer emotes manager

* Make the output message for the steam command editable

* Remove old unused processors

* Catch exceptions for twitch team API requests

* Create separate chat settings for third party emotes

* Add an option to edit overlay instances from an effect

* Add a confirmation modal for hotkey deletion

* Add a day selection option to the time range restriction. (#1536)

* Fix double output when using the currency command

* Fix custom scripts + increment version

* Streamlabs Chatbot quotes importer (#1520)

+++ gui/app/controllers/hotkeys.controller.js
@@ -35,7 +35,18 @@
                             hotkeyService.updateHotkey(hotkey);
                             break;
                         case "delete":
-                            hotkeyService.deleteHotkey(hotkey);
+                            utilityService
+                                .showConfirmationModal({
+                                    title: "Delete Hotkey",
+                                    question: `Are you sure you want to delete the Hotkey "${hotkey.name}"?`,
+                                    confirmLabel: "Delete",
+                                    confirmBtnType: "btn-danger"
+                                })
+                                .then(confirmed => {
+                                    if (confirmed) {
+                                        hotkeyService.deleteHotkey(hotkey);
+                                    }
+                                });
                             break;
                         }
                     }
